+++ src/Commands/DomainDtoMakeCommand.php
@@ -2,9 +2,7 @@
 
 namespace Lunarstorm\LaravelDDD\Commands;
 
-use Symfony\Component\Console\Input\InputArgument;
-
-class MakeDTO extends DomainGeneratorCommand
+class DomainDtoMakeCommand extends DomainGeneratorCommand
 {
     protected $name = 'ddd:dto';
 
@@ -17,17 +15,15 @@ class MakeDTO extends DomainGeneratorCommand
 
     protected $type = 'Data Transfer Object';
 
-    protected function getArguments()
+    protected function configure()
     {
-        return [
-            ...parent::getArguments(),
+        $this->setAliases([
+            'ddd:data-transfer-object',
+            'ddd:datatransferobject',
+            'ddd:data',
+        ]);
 
-            new InputArgument(
-                'name',
-                InputArgument::REQUIRED,
-                'The name of the DTO',
-            ),
-        ];
+        parent::configure();
     }
 
     protected function getStub()
@@ -37,7 +33,7 @@ class MakeDTO extends DomainGeneratorCommand
 
     protected function getRelativeDomainNamespace(): string
     {
-        return config('ddd.namespaces.data_transfer_objects', 'Data');
+        return config('ddd.namespaces.data_transfer_object', 'Data');
     }
 
     protected function preparePlaceholders(): array