+++ src/Commands/DomainListenerMakeCommand.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Lunarstorm\LaravelDDD\Commands;
+
+use Illuminate\Foundation\Console\ListenerMakeCommand;
+use Lunarstorm\LaravelDDD\Commands\Concerns\ResolvesDomainFromInput;
+
+class DomainListenerMakeCommand extends ListenerMakeCommand
+{
+    use ResolvesDomainFromInput;
+
+    protected $name = 'ddd:listener';
+}