+++ src/Commands/DomainScopeMakeCommand.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Lunarstorm\LaravelDDD\Commands;
+
+use Illuminate\Foundation\Console\ScopeMakeCommand;
+use Lunarstorm\LaravelDDD\Commands\Concerns\ResolvesDomainFromInput;
+
+class DomainScopeMakeCommand extends ScopeMakeCommand
+{
+    use ResolvesDomainFromInput;
+
+    protected $name = 'ddd:scope';
+}